Used a single package for new/old embedding
