+++ packages/firebase_performance/android/src/main/java/io/flutter/plugins/firebaseperformance/FlutterHttpMetric.java
@@ -9,9 +9,11 @@ import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 
 public class FlutterHttpMetric implements MethodChannel.MethodCallHandler {
+  private final FirebasePerformanceHandler handler;
   private final HttpMetric httpMetric;
 
-  FlutterHttpMetric(final HttpMetric metric) {
+  FlutterHttpMetric(FirebasePerformanceHandler handler, final HttpMetric metric) {
+    this.handler = handler;
     this.httpMetric = metric;
   }
 
@@ -60,7 +62,7 @@ public class FlutterHttpMetric implements MethodChannel.MethodCallHandler {
     httpMetric.stop();
 
     final Integer handle = call.argument("handle");
-    FirebasePerformancePlugin.removeHandler(handle);
+    handler.removeHandler(handle);
 
     result.success(null);
   }
